fixed warn
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@4915 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/utils/probe/term-iostream/termiostreamdisplay.cpp
@@ -301,7 +301,7 @@ TermIostreamDisplay::printRegistryData( const orcacm::RegistryHierarchicalData1&
             int selection = atoi( userInput.c_str() );
             cout<<"you picked "<<selection<<endl;
             // check
-            if ( selection>=0 && selection<data.platforms.size() ) {
+            if ( selection>=0 && selection<(int)data.platforms.size() ) {
                 browser_->choosePick( selection );
                 break;
             }
